// src/shared/api/model/requestUserRegister.ts
/**
 * Generated by orval v7.1.1 🍺
 * Do not edit manually.
 * Anime ON API
 * OpenAPI spec version: ##VERSION## DEV
 */

export interface RequestUserRegister {
  /**
   * @maxLength 254
   * @nullable
   */
  email: string | null;
  /**
   * @minLength 1
   * @maxLength 128
   */
  password: string;
  /** @minLength 1 */
  password_repeat: string;
  /**
   * Required. 150 characters or fewer. Letters, digits and @/./+/-/_ only.
   * @minLength 1
   * @maxLength 150
   * @pattern ^[\w.@+-]+$
   */
  username: string;
}